backend/requirements.txt
asgiref==3.3.1
astroid==2.5.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
defusedxml==0.7.1
Django==3.1.7
django-braces==1.14.0
django-cors-headers==3.7.0
django-oauth-toolkit==1.4.0
django-rest-framework-social-oauth2==1.1.0
django-webpack-loader==0.7.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
dynaconf==3.1.2
idna==2.10
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
oauthlib==3.1.0
pycparser==2.20
PyJWT==2.0.1
pylint==2.7.2
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.3
wrapt==1.12.1
